requirements.txt
asgiref==3.5.2
Django==4.0.4
gunicorn==20.1.0
psycopg2==2.8.6
sqlparse==0.4.2


dj-database-url==1.0.0

django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
python-dotenv==0.21.0
pytz==2022.5

whitenoise==6.2.0